FEAT: Add Qt 6 support

+++ src/murmur/DBus.h
@@ -95,7 +95,7 @@ Q_DECLARE_METATYPE(RegisteredPlayer)
 Q_DECLARE_METATYPE(QList< RegisteredPlayer >)
 
 struct LogEntry {
-	unsigned int timestamp;
+	qint64 timestamp;
 	QString txt;
 	LogEntry();
 	LogEntry(const ServerDB::LogRecord &);
